+++ src/main/java/calculator/view/InputView.java
@@ -1,4 +1,4 @@
-package view;
+package calculator.view;
 
 import java.util.Scanner;
