+++ src/main/java/calculator/view/OutputView.java
@@ -1,4 +1,4 @@
-package view;
+package calculator.view;
 
 public class OutputView {
     public static void printResult(int result) {
